// src/hooks/use-gracefully/types.ts
export interface GracefulState {
  /**
   * @example
   * const { is, breakpoints, devices, window } = useGracefully();
   * console.log(breakpoints); // { sm: { min: undefined, max: '500px' }, md: { min: '501px', max: '1000px' }, lg: { min: '1001px', max: undefined } }
   *
   * @description
   * Returns currently configured breakpoints
   *
   * @returns {Object} breakpoints - currently configured breakpoints
   */
  breakpoints: Breakpoints;
  /**
   * @example
   * const { is, breakpoints, devices, window } = useGracefully();
   * console.log(devices); // ['mobile', 'android']
   *
   * @description
   * Returns list of currently active devices
   *
   * @returns {Object} devices - currently active devices
   */
  devices: Device[];
  /**
   * @example
   * const { is, breakpoints, devices, window } = useGracefully();
   * console.log(window); // { width: '1000px', height: '1000px', orientation: 'portrait' }
   *
   * @description
   * Returns current window sizes and orientation
   *
   * @returns {Object} windowState - current window sizes and orientation
   */
  window: WindowState;
  /**
   * @example
   * const { is } = useGracefully();
   * const isMobile = is.mobile();
   * const isAboveMedium = is.above.breakpoint('md');
   * const isWindowHeightBelow500px = is.below.window.height('500px');
   * const isLandscape = is.current.window.landscape();
   *
   * @description
   * Collection of functions used to get details about device, breapoints and/or window sizes or orientation
   *
   * @returns {Object} is - collection of functions to get details about device, breakpoint and/or window sizes or orientation
   */
  is: GracefulStateIsFunctions;
}